+++ app/components/index.ts
@@ -1,3 +1,6 @@
-export { Callout } from './callout';
-export { YouTubeComponent } from './youtube';
-export type { CalloutProps } from './callout'; 
+export { Callout } from './mdx-blocks/Callout';
+export { YouTubeComponent } from './media/YouTube';
+export type { CalloutProps } from './mdx-blocks/Callout';
+export { default as ClientMDX } from "./utils/ClientMDX";
+export { default as ErrorBoundary } from "./utils/ErrorBoundary";
+export { default as P5Container } from "./utils/P5Container";
