+++ app/components/mdx-blocks/Caption.tsx
