+++ app/components/mdx-blocks/CollapseMetadata.tsx
