+++ app/components/mdx-blocks/HeroBlock.tsx
