+++ app/components/mdx-blocks/MediaSection.tsx
