+++ app/components/mdx-blocks/TechnicalDetails.tsx
