+++ app/components/surreal-systems/GalleryOfLies.tsx
