+++ app/layout.tsx
@@ -1,11 +1,11 @@
 import "./global.css";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import { Navbar } from "./components/nav";
+import { Navbar } from "./components/layout/Nav";
 import { Analytics } from "@vercel/analytics/react";
 import { SpeedInsights } from "@vercel/speed-insights/next";
-import Footer from "./components/footer";
-import { ThemeProvider } from "./components/theme-switch";
+import Footer from "./components/layout/Footer";
+import { ThemeProvider } from "./components/layout/ThemeSwitch";
 import { metaData } from "./config";
 import "katex/dist/katex.min.css";
