+++ content/projects/gallery-of-lies.mdx
@@ -15,6 +15,6 @@ collapseMode: false
 clusterOrder: 2
 ---
 
-import GalleryOfLies from "@/app/components/GalleryOfLies"
+import GalleryOfLies from "@/app/components/surreal-systems/GalleryOfLies"
 
 <GalleryOfLies mode="gallery" />
