+++ content/projects/rain.mdx
@@ -15,20 +15,20 @@ clusterOrder: 2
 ---
 
 
-import { CaptionComponent as Caption } from '@/app/components/caption';
+import { CaptionComponent as Caption } from '@/app/components/mdx-blocks/Caption';
 import ZoomImage from "@/components/ZoomImage";
 
 <ProjectOverview>
-*Rain* isn’t structured — it’s **carved**.
+*Rain* isn't structured — it's **carved**.
 
 Like the Cantor ternary set, it begins whole.  
 Then something is taken away. Then more. Then again.  
-What’s left is silence in the shape of a pattern.
+What's left is silence in the shape of a pattern.
 
 This piece removes its own center. Recursively.  
 Rhythms drop out. Frequencies fall into negative space.
 
-It’s not music about rain. This is rain rendered algorithmic: broken into recursive gaps, twitching at the edge of rhythm.
+It's not music about rain. This is rain rendered algorithmic: broken into recursive gaps, twitching at the edge of rhythm.
 
 <Image
   src="/photos/cantor.png"
