+++ content/projects/spatial-synthesizer.mdx
@@ -16,13 +16,13 @@ collapseMode: true
 clusterOrder: 6
 ---
 
-import ProjectOverview from '@/app/components/ProjectOverview'
-import MediaSection from '@/app/components/MediaSection'
-import MediaItem from '@/app/components/MediaItem'
-import SpatialSynthesizerSketch from '@/app/components/SpatialSynthesizerSketch'
-import CallToAction from '@/app/components/CallToAction'
-import FieldNote from '@/app/components/FieldNote'
-import CollapseMetadata from '@/app/components/CollapseMetadata'
+import ProjectOverview from '@/app/components/mdx-blocks/ProjectOverview'
+import MediaSection from '@/app/components/mdx-blocks/MediaSection'
+import MediaItem from '@/app/components/mdx-blocks/MediaItem'
+import SpatialSynthesizerSketch from '@/app/components/surreal-systems/SpatialSynthesizerSketch'
+import CallToAction from '@/app/components/mdx-blocks/CallToAction'
+import FieldNote from '@/app/components/mdx-blocks/FieldNote'
+import CollapseMetadata from '@/app/components/mdx-blocks/CollapseMetadata'
 
 <ProjectOverview>
  What if geometry behaved like tone?
